+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 - Add `dump_context` global/local setting to include `context` in `Task#to_h`
 - Add `Result#success!` for annotating successful results
 - Add `started_at` and `ended_at` to runtime middleware payload
+- Add `keys`, `values`, `each`, `each_key`, `each_value` methods to context
 
 ### Changed
 - Move `faults.unspecified` locale key to `reasons.unspecified`